:truck: move race abilities to race folder
